% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clip_partial_month}
\alias{clip_partial_month}
\title{Given a data frame with a \code{date} column, remove any partial month at the
beginning or end.}
\usage{
clip_partial_month(df)
}
\description{
Given a data frame with a \code{date} column, remove any partial month at the
beginning or end.
}
