% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{daily_max_power_histogram}
\alias{daily_max_power_histogram}
\title{Histogram of daily maximum power generation}
\usage{
daily_max_power_histogram()
}
\description{
Histogram of daily maximum power generation
}
