% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{daily_usage_vs_generation}
\alias{daily_usage_vs_generation}
\title{Actual daily usage vs actual and estimated generation}
\usage{
daily_usage_vs_generation()
}
\description{
Actual daily usage vs actual and estimated generation
}
