% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fns.R
\name{energy_generation}
\alias{energy_generation}
\title{Get total energy generation (kWh) between two dates}
\usage{
energy_generation(start_date, end_date, ...)
}
\arguments{
\item{start_date}{The first date to count}

\item{end_date}{The date \emph{after} the last date (open interval)}
}
\description{
Get total energy generation (kWh) between two dates
}
