% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fns.R
\name{energy_summary}
\alias{energy_summary}
\title{Get total energy generation and consumption for all available dates}
\usage{
energy_summary()
}
\description{
Get total energy generation and consumption for all available dates
}
