% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{energy_table}
\alias{energy_table}
\title{Get a \code{dplyr} table backed by the Energy table}
\usage{
energy_table(envir = parent.frame())
}
\value{
A table
}
\description{
Get a \code{dplyr} table backed by the Energy table
}
