% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_requests.R
\name{get_daily_energy}
\alias{get_daily_energy}
\title{Get daily energy data}
\usage{
get_daily_energy(start_date, end_date)
}
\arguments{
\item{start_date}{string or DateTime to start}

\item{end_date}{string or Date to end, must be <= now()}
}
\value{
A data frame with columns \code{date}, \code{value}, \code{unit},
\code{year}, \code{month}, \code{day}, \code{time}
}
\description{
Get daily energy data
}
