% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_requests.R
\name{get_energy_details}
\alias{get_energy_details}
\title{Get detailed energy data}
\usage{
get_energy_details(
  start_time,
  end_time,
  time_unit = c("QUARTER_OF_AN_HOUR", "HOUR", "DAY", "WEEK", "MONTH", "YEAR ")
)
}
\arguments{
\item{start_time}{string or DateTime to start}

\item{end_time}{string or DateTime to end, must be <= now()}

\item{time_unit}{Granularity to report}
}
\value{
A data frame with columns \code{meter}, \code{date}, \code{value}, \code{unit}
}
\description{
Get detailed energy data
}
