% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_requests.R
\name{get_power}
\alias{get_power}
\title{Get detailed power data}
\usage{
get_power(start_time, end_time)
}
\arguments{
\item{start_time}{string or DateTime to start}

\item{end_time}{string or DateTime to end, must be <= now()}
}
\value{
A data frame with columns \code{date}, \code{value}, \code{unit},
\code{year}, \code{month}, \code{day}, \code{time}
}
\description{
Get detailed power data
}
