% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{init_db}
\alias{init_db}
\title{One-time database initialization}
\usage{
init_db(start_date)
}
\arguments{
\item{start_date}{Earliest date to get data from the API}
}
\description{
One-time database initialization
}
