% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{local_db}
\alias{local_db}
\title{Get a database connection that will be correctly shut down
when it goes out of scope in the provided environment}
\usage{
local_db(envir = parent.frame())
}
\arguments{
\item{envir}{Environment to monitor for scope}
}
\description{
Get a database connection that will be correctly shut down
when it goes out of scope in the provided environment
}
