% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_date}
\alias{make_date}
\title{Convert input date from string if needed}
\usage{
make_date(d)
}
\arguments{
\item{d}{A Date or string}
}
\value{
A Date
}
\description{
Convert input date from string if needed
}
