% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{monthly_discrepancy_chart}
\alias{monthly_discrepancy_chart}
\title{Difference between monthly energy (kWh) and predicted
as a percent of predicted.}
\usage{
monthly_discrepancy_chart()
}
\description{
Difference between monthly energy (kWh) and predicted
as a percent of predicted.
}
