% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{monthly_predictions}
\alias{monthly_predictions}
\title{Read the PV2 predictions of monthly energy in kWh}
\usage{
monthly_predictions()
}
\value{
A data.frame with columns Month, kWh
}
\description{
Read the PV2 predictions of monthly energy in kWh
}
