% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{power_chart}
\alias{power_chart}
\title{15-minute power (watts) for a month, by day}
\usage{
power_chart(start_date = today())
}
\arguments{
\item{start_date}{A date within the desired month}
}
\description{
15-minute power (watts) for a month, by day
}
