% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{power_table}
\alias{power_table}
\title{Get a \code{dplyr} table backed by the Power table}
\usage{
power_table(envir = parent.frame())
}
\value{
A table
}
\description{
Get a \code{dplyr} table backed by the Power table
}
