% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_requests.R
\name{solaredge_secrets}
\alias{solaredge_secrets}
\title{Get the SolarEdge key and site id from the environment}
\usage{
solaredge_secrets()
}
\value{
A list with \code{site_id} and \code{key} members
}
\description{
Get the SolarEdge key and site id from the environment
}
