% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{update_energy}
\alias{update_energy}
\title{Update the energy table with the latest}
\usage{
update_energy()
}
\value{
A data frame with columns \code{date}, \code{value}, \code{unit},
\code{year}, \code{month}, \code{day}, \code{time}
}
\description{
Update the energy table with the latest
}
