% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{update_power}
\alias{update_power}
\title{Update the power table with the latest at 15-minute intervals}
\usage{
update_power()
}
\value{
A data frame with columns \code{date}, \code{value}, \code{unit},
\code{year}, \code{month}, \code{day}, \code{time}
}
\description{
Update the power table with the latest at 15-minute intervals
}
